environments/verihop/pyproject.toml
[project]
name = "verihop"
description = "VeriHop — modular multi-hop multimodal RLVR with procedural scenes and dense rubrics"
tags = ["multimodal", "multi-turn", "tools", "train", "eval", "rlvr"]
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "verifiers>=0.1.11",
    "pillow>=11.0.0",
    "datasets>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["verihop"]

[tool.verifiers.eval]
num_examples = 8
rollouts_per_example = 1
